Enable tagging API gateway (#148)
This adds the ability to tag the underlying resources of the API gateway component.
Closes #139

I cleaned up some dead code that was a leftover from extracting the API gateway component out of awsx. The resource aws:apigateway:x:API was not exposed in this provider and thus not included in the SDKs.