chore: fix renovate
PR #180 updated the Makefile to add these targets. It mentions basing it
on pulumi-eks, but it looks like the configuration for downloading pulumictl is not used by
any other repo and is causing problems. All the others just assume that
it is available so maybe something changed recently with our renovate
config to allow for that assumption.

re #184