Onboard aws-apigateway to ci-mgmt (#189)
This PR contains the work required to onboard this provider with
ci-mgmt.

Changes made:

- Moved the nodejs provider code to `aws-apigateway/` (to align with
other node providers like awsx)
- Created shim code in `provider/cmd/pulumi-gen-aws-apigateway` which
shells out to `pulumi package gen-sdk` (to align with how other
providers generate SDKs and schema)
- pulumi-gen-aws-apigateway now generates a schema.json file in
`provider/cmd/pulumi-resource-aws-apigateway/schema.json`
- Added new files/folders from ci-mgmt to .gitignore

To ensure this shim and repo structure works, #190 contains the
generated ci-mgmt generated workflow files and is passing. An alpha
version is tagged to the HEAD of this branch and also passes
verifications.